MAINT: Make numpy.linalg.linalg private
