+++ numpy/linalg/_linalg.pyi
